fix: last things to init working

+++ BackendMegaPet/User/Resources/UserResource.cs
@@ -5,7 +5,7 @@ public class UserResource
     public int Id { get; set; }
     public string name { get; set; }
     public string lastName { get; set; }
-    public long phone { get; set; }
+    public int phone { get; set; }
     public string email { get; set; }
     public string image { get; set; }
     public string password { get; set; }
